Restructured the Simulator package.
